# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Project Meridian API
  version: v1
  description: Product lifecycle management platform API.
paths:
...
  /api/v1/products:
    get:
      operationId: products_list
      parameters:
      - in: query
        name: brand_code
        schema:
          type: string
      - in: query
        name: category_code
        schema:
          type: string
      - in: query
        name: channel_code
        schema:
          type: string
      - in: query
        name: external_id
        schema:
          type: string
      - in: query
        name: lifecycle_status
        schema:
          type: string
      - in: query
        name: owner_public_id
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
      - in: query
        name: sku_code
        schema:
          type: string
      tags:
      - products
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSearchPage'
          description: ''
  /api/v1/products/{public_id}:
    get:
      operationId: products_retrieve
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - products
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
          description: ''
...
components:
  schemas:
...
    ExternalBinding:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        source_system:
          type: string
        object_type:
          type: string
        external_id:
          type: string
        binding_status:
          type: string
      required:
      - binding_status
      - external_id
      - object_type
      - public_id
      - source_system
...
    ProductDetail:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        business_no:
          type: string
        name:
          type: string
        lifecycle_status:
          type: string
        brand_code:
          type: string
        category_code:
          type: string
        formula_summary:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ProductVersionSummary'
        external_bindings:
          type: array
          items:
            $ref: '#/components/schemas/ExternalBinding'
      required:
      - brand_code
      - business_no
      - category_code
      - external_bindings
      - lifecycle_status
      - name
      - public_id
      - versions
...
    ProductSearchPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
        page:
          type: integer
        page_size:
          type: integer
        count:
          type: integer
      required:
      - count
      - items
      - page
      - page_size
    ProductSkuSummary:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        sku_code:
          type: string
        name:
          type: string
        specification:
          type: string
        barcode:
          type: string
        channels:
          type: array
          items: {}
      required:
      - name
      - public_id
      - sku_code
      - specification
    ProductSummary:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        business_no:
          type: string
        name:
          type: string
        lifecycle_status:
          type: string
        formula_summary:
          type: string
      required:
      - business_no
      - lifecycle_status
      - name
      - public_id
    ProductVersionSummary:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        version_code:
          type: string
        version_name:
          type: string
        status:
          type: string
        skus:
          type: array
          items:
            $ref: '#/components/schemas/ProductSkuSummary'
      required:
      - public_id
      - skus
      - status
      - version_code
      - version_name
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid